backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "amphoreus"
version = "0.1.0"
description = "AI content operations platform — bitter-lesson-pilled ghostwriting pipeline"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-multipart>=0.0.9",

    # LLM providers
    "anthropic>=0.40.0",
    "google-genai>=1.0.0",
    "openai>=1.50.0",
    "instructor>=1.5.0",

    # Data stores
    "supabase>=2.9.0",
    "pinecone>=8.0.0",

    # Audio capture + transcription
    "faster-whisper>=1.0.0",          # local fallback path
    "deepgram-sdk>=3.7.0",            # Nova-3 streaming (primary path)
    "sounddevice>=0.5.0",
    "numpy>=1.26.0",

    # Sandbox
    "e2b-code-interpreter>=1.0.0",

    # Image processing
    "Pillow>=10.0.0",

    # Utilities
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "apify-client>=2.5.0",
    "tiktoken>=0.7.0",
    "openpyxl>=3.1.0",
    "pyjwt[crypto]>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
